requirements-dev.txt
# Runtime requirements
--requirement requirements.txt

setuptools>=68.2.2
build>=1.0.3
twine>=4.0.2
dataclasses>=0.6
pillow>=10.1.0

# Linting/Tooling
black>=23.11.0
isort>=5.12.0
mypy>=1.6.1
pre-commit>=3.5.0
ruff>=0.1.5
pyright>=1.1.335
pathspec>==0.11.2
platformdirs>=4.0.0
tomli>=2.0.1
typing-extensions>=4.4.0
click>=8.0.3

# Testing
pytest>=7.4.3
codecov>=2.1.13
pytest-cov>=4.1.0
pytest-benchmark>=4.0.0

# Documentation
mkdocs>=1.5.3
mkdocstrings>=0.23.0
mkdocstrings-python>=1.7.3
mkdocs-material>=9.4.8
Pygments>=2.16.1

#frontend
# tkinter>= is in python3

#devel
# cryptography>=41.0.7
# pycrypto>=2.6.1
pycryptodome>=3.19.0
